fixing decrement changes

+++ assignment_5/a4_starter/BatchCommand.java
@@ -1,12 +1,12 @@
 import io.atomix.copycat.Command;
 
 import java.util.Map;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class BatchCommand implements Command {
-    Map<String, AtomicInteger> _changes;
+    Map<String, AtomicLong> _changes;
 
-    public BatchCommand(Map<String, AtomicInteger> changes) {
+    public BatchCommand(Map<String, AtomicLong> changes) {
         _changes = changes;
     }
 }
